% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tone_curve.R
\name{mod_tone_curve_ui}
\alias{mod_tone_curve_ui}
\alias{mod_tone_curve_server}
\title{mod_tone_curve_ui and mod_tone_curve_server}
\usage{
mod_tone_curve_ui(id)

mod_tone_curve_server(input, output, session, tone_curve_points,
  tone_curve)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
